Rename `util` to `common` and shuffle some stuff around
