+++ src/poprf.rs
@@ -19,14 +19,13 @@ use generic_array::typenum::{IsLess, IsLessOrEqual, Unsigned, U256};
 use generic_array::GenericArray;
 use rand_core::{CryptoRng, RngCore};
 
-use crate::group::STR_HASH_TO_SCALAR;
-#[cfg(feature = "serde")]
-use crate::serialization::serde::{Element, Scalar};
-use crate::util::{
+use crate::common::{
     create_context_string, derive_keypair, deterministic_blind_unchecked, generate_proof, i2osp_2,
     verify_proof, BlindedElement, EvaluationElement, Mode, PreparedEvaluationElement, Proof,
-    STR_FINALIZE, STR_INFO,
+    STR_FINALIZE, STR_HASH_TO_SCALAR, STR_INFO,
 };
+#[cfg(feature = "serde")]
+use crate::serialization::serde::{Element, Scalar};
 use crate::{CipherSuite, Error, Group, Result};
 
 ////////////////////////////
@@ -772,7 +771,7 @@ mod tests {
     use rand::rngs::OsRng;
 
     use super::*;
-    use crate::group::STR_HASH_TO_GROUP;
+    use crate::common::STR_HASH_TO_GROUP;
     use crate::Group;
 
     fn prf<CS: CipherSuite>(
